NoteMasterAI/requirements.txt
PyQt5
pdfplumber
pdf2image
opencv-python
google-cloud-vision
google-generativeai
requests
numpy
ultralytics
Pillow
python-dotenv
